test(fab): migrate translucent
